requirements.txt
dj-database-url==0.4.0
Django==1.10
django-robokassa==1.3
gunicorn==19.4.5
psycopg2==2.6.2
whitenoise==2.0.6